Move sign up page to security app
