+++ security/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
 
-from .views import SignInView
+from .views import SignUpWizard, SignInView
 
 urlpatterns = [
+    path("sign-up/", SignUpWizard.as_view(), name="sign-up"),
     path("sign-in/", SignInView.as_view(), name="sign-in"),
 ]
