+++ talent/urls.py
@@ -7,12 +7,10 @@ from django.contrib.auth.views import (
     PasswordResetCompleteView,
 )
 
-from .views import SignUpWizard, SignInView, ResetPasswordView
+from .views import ResetPasswordView
 from . import views
 
 urlpatterns = [
-    path("sign-up/", SignUpWizard.as_view(), name="sign-up"),
-    path("sign-in/", SignInView.as_view(), name="sign-in"),
     path(
         "reset-password/",
         ResetPasswordView.as_view(),
